Amend code structure and format
